refactor: make tests independent
